requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
